requirements.txt
azure-core==1.27.0
azure-devops==7.1.0b3
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
html-diff==0.4.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.6.3
urllib3==2.0.3
